@@ -2,9 +2,15 @@
   "name": "threads-clone",
   "description": "Demo App with Routing built-in (recommended)",
   "engines": {
-    "node": ">=15.0.0"
+    "node": "^18.17.0 || ^20.3.0 || >=21.0.0"
   },
+  "engines-annotation": "Mostly required by sharp which needs a Node-API v9 compatible runtime",
   "private": true,
+  "trustedDependencies": [
+    "sharp"
+  ],
+  "trustedDependencies-annotation": "Needed for bun to allow running install scripts",
+  "type": "module",
   "scripts": {
     "build": "qwik build",
     "build.client": "vite build",
@@ -22,35 +28,35 @@
     "qwik": "qwik"
   },
   "devDependencies": {
-    "@builder.io/qwik": "^1.2.15",
-    "@builder.io/qwik-city": "^1.2.15",
-    "@types/eslint": "8.44.6",
-    "@types/node": "^20.8.10",
-    "@types/react": "18.2.33",
-    "@types/react-dom": "18.2.14",
-    "@typescript-eslint/eslint-plugin": "6.9.1",
-    "@typescript-eslint/parser": "6.9.1",
-    "autoprefixer": "^10.4.16",
-    "eslint": "8.52.0",
-    "eslint-plugin-qwik": "^1.2.15",
-    "postcss": "^8.4.31",
-    "prettier": "3.0.3",
-    "tailwindcss": "^3.3.5",
-    "typescript": "5.2.2",
-    "undici": "5.27.0",
-    "vercel": "^32.5.0",
-    "vite": "4.5.0",
-    "vite-tsconfig-paths": "4.2.1"
+    "@builder.io/qwik": "^1.5.1",
+    "@builder.io/qwik-city": "^1.5.1",
+    "@types/eslint": "8.56.5",
+    "@types/node": "^20.11.25",
+    "@types/react": "18.2.64",
+    "@types/react-dom": "18.2.21",
+    "@typescript-eslint/eslint-plugin": "7.1.1",
+    "@typescript-eslint/parser": "7.1.1",
+    "autoprefixer": "^10.4.18",
+    "eslint": "8.57.0",
+    "eslint-plugin-qwik": "^1.5.1",
+    "postcss": "^8.4.35",
+    "prettier": "3.2.5",
+    "tailwindcss": "^3.4.1",
+    "typescript": "5.4.2",
+    "undici": "6.7.1",
+    "vercel": "^33.5.5",
+    "vite": "5.1.5",
+    "vite-tsconfig-paths": "4.3.1"
   },
   "dependencies": {
-    "@builder.io/qwik-react": "0.5.0",
-    "@cloudinary/url-gen": "^1.12.0",
-    "@unpic/qwik": "^0.0.29",
+    "@builder.io/qwik-react": "0.5.4",
+    "@cloudinary/url-gen": "^1.19.0",
+    "@unpic/qwik": "^0.0.44",
     "dayjs": "^1.11.10",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "sharp": "^0.32.6",
-    "swiper": "^11.0.3"
+    "sharp": "^0.33.2",
+    "swiper": "^11.0.7"
   },
   "volta": {
     "node": "18.18.0"
